requirements_mac.txt
filetype==1.2.0
idna==3.4
loguru==0.6.0
Pillow==9.4.0
pyperclip==1.8.2
python-tkdnd==0.2.1
requests==2.28.2
tkinterdnd2==0.3.0
ttkbootstrap==1.10.1
ttkwidgets==0.13.0
urllib3==1.26.14
pasteboard==0.3.3
chardet==5.1.0
py2app==0.28.5
